+++ plugins/woocommerce-admin/packages/experimental/src/index.js
@@ -38,3 +38,5 @@ export { ExperimentalList as List } from './experimental-list/experimental-list'
 export { ExperimentalCollapsibleList as CollapsibleList } from './experimental-list/collapsible-list';
 export { TaskItem } from './experimental-list/task-item';
 export * from './inbox-note';
+
+export * from './vertical-css-transition';