+++ plugins/woocommerce-admin/packages/experimental/CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+-   Add new VerticalCSSTransition component for handling height transitions. #7203
+-   Update TaskItem to make use of the VerticalCSSTransition. #7203
+
 # 1.3.0
 
 -   Remove the use of Dashicons and replace with @wordpress/icons or gridicons #7020